Conceptual change to Resolution Analysis
Rather than now doing resolution as yearly running means over a window
of X years, doing snapshots where we’re extracting the value of that
scale every x years.  For instance, now instead “resolution” being the
100-year smoothed (“t.100”, running mean) for all 1161 years at a site,
we take the 100-year mean value (still “t.100), every 100 years.

This makes it more similar to the values we get if we resample a forest
every [100 years] and figure out what the average annual increment must
be.

I think this is a more accurate way to portray resolution and more
consistent with the extent analyses than running means.  The original
way ended up with the curves based on lower resolution still having the
same n, which is misleading.