Bug Fixes: Years, stability
Changed the selection of years by resolution to start with 2010 and go
back towards 850.  This makes sure that we will always get the most
recent years in the data set.  Going the other way, we would stop at
1950, which gave us problematically truncated temp & CO2 distributions

Also did a few tweaks to get LPJ & CLM to be more stable