Modify GAMM structure
Instead of using fixed intercepts for Site/PlotID, added a effect that
allows for a spatially & temporally dynamic factor that we think
explains most of the variation among sites & non-climatic variation in
NPP through time.

I did play around with making time a smoother, but it did nonsensical
things, so I decided to prescribe a log() effect.

*** This model structure now extends even to tree rings where we now
use raw ring width as the response and DBH as the predictor ***