Update GAMM structure
linear “Time” (usually AGB) effect for most models.  This was
determined by some quick AIC comparisons and visual comparisons between
AGB & NPP across sites.  “Time” for tree ring width models (TreeRingRW)
is the log of DBH.  This was chosen based on AIC for the “baseline”
model as well as to get something a little closer to the negative
exponential curve typically used in conservative detrending.