Figures updated; CO2 weight standardized
Updated code for figures to only show temp, precip & CO2 sensitivity
(ignoring Year & Time).

*** Also: To keep the past in the “drivers of NPP wiggling through
time” figures from showing as being completely carbon limited prior to
the “0” point, the CO2 effect was shifted based on the mean CO2 effect
at the start of the run.  For example, a simulation based on 1985-2010
will have the CO2 weight start as 0 in 1985. This offset is then added
to the rest of the series (so 2010 is even more CO2 fertilized).  Not
doing this standardization causes really wonky things because unlike
precip & temp, CO2 doesn’t generally fluctuate around its mean through
time.  This means it’s “break even point”, particularly for shorter
runs is about 3 years in the mid-90s, which is not how it actually
works in nature.