GAM conceptual change
For Models, this is mostly a terminology change to be consistent with
the manuscript, but now we’re looking at NPP (or Ring Width) as a
function of climate (temp/precip), CO2, Biomass (an indicator of how
much NPP you theoretically could produce), and Time (succession).

For the models Time = simulation year, basically assuming steady state
state at the beginning.  This is deliberately a more rigid and
prescribed shape than the climate sensitivity to try and remove
endogenous low-frequency patterns. If a model or NPP or tree growth is
at steady state, this term should be non-significant and all wiggles
would be due to climate, and possibly a directional trend ascribed to
Biomass.  It might be possible to compare Time curves to get an idea of
if/when models get away from steady state.  For NPP, a forest-level
estimate, time is the stand age.  While we’ve restricted analyses to
the past 30 years to try and remove succession, there may be still some
slight differences because of different ages of plots.  This effect
should be greatest for tree ring width, which we’re acknowledging off
the bat is not at equilibrium since the tree is growing rapidly;
particularly for the non-ITRDB samples.